Change Category model to resource.categories
