+++ resources/categories/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from resources.categories.models import Category
+
+# Register your models here.
+admin.site.register(Category)
