+++ resources/days/admin.py
@@ -1,6 +1,5 @@
 from django.contrib import admin
-from .models import Category, Day
+from .models import Day
 
 # Register your models here.
 admin.site.register(Day)
-admin.site.register(Category)
